+++ Changes.md
@@ -94,6 +94,7 @@ API
 - ShufflePlugValueWidget : Widgets for the `source` and `destination` plugs can now be customised using standard `plugValueWidget:type` metadata.
 - ImageTestCase : in `assertImageEqual` function, maxDifference may now be a tuple, to specify an asymmetric range.
 - Editor : Added `Settings` class, which should be used to store settings for subclasses. See LightEditor and ImageInspector for examples.
+- DeepPixelAccessor : Added utility class for accessing deep samples while abstracting away the underlying tile storage.
 
 Breaking Changes
 ----------------